unified-system-v1/requirements.txt
# Core Dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
pyzmq>=25.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Logging
python-json-logger>=2.0.0
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Utilities
click>=8.1.0
tabulate>=0.9.0
tqdm>=4.66.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# AI/ML (Optional - for local LLM support)
# transformers>=4.35.0
# torch>=2.0.0
# accelerate>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# Cloud SDKs (Optional)
# boto3>=1.28.0  # AWS
# azure-storage-blob>=12.19.0  # Azure
# google-cloud-storage>=2.10.0  # GCP# Additional dependencies from common modules
redis>=4.0.0
nats-py>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
alembic>=1.8.0
passlib>=1.7.0
python-jose>=3.3.0
httpx>=0.23.0
uvicorn>=0.18.0
fastapi>=0.85.0
pydantic>=1.10.0
